feat(llm): add the lns upsell banner on the wallet screen
