+++ apps/ledger-live-mobile/src/newArch/features/LNSUpsell/components/LNSUpsellBanner/types.ts
@@ -1,9 +1,4 @@
-import { LlmNanoSUpsellBannersConfig } from "@ledgerhq/types-live/lib/lnsUpsell";
-
-export type LNSBannerLocation = Extract<
-  "manager" | "accounts" | "notification_center" | "wallet",
-  keyof LlmNanoSUpsellBannersConfig
->;
+export type { LNSBannerLocation } from "../../types";
 
 export type LNSBannerModel = {
   isShown: boolean;
