+++ apps/ledger-live-mobile/src/newArch/features/LNSUpsell/components/LNSUpsellBanner/useLNSUpsellBannerModel.ts
@@ -1,29 +1,10 @@
 import { Linking } from "react-native";
-import { useSelector } from "react-redux";
-import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { track } from "~/analytics";
-import { accountsSelector } from "~/reducers/accounts";
-import {
-  knownDeviceModelIdsSelector,
-  personalizedRecommendationsEnabledSelector,
-} from "~/reducers/settings";
 import type { LNSBannerLocation, LNSBannerModel } from "./types";
+import { useLNSUpsellBannerState } from "../../hooks/useLNSUpsellBannerState";
 
 export function useLNSUpsellBannerModel(location: LNSBannerLocation): LNSBannerModel {
-  const isOptIn = useSelector(personalizedRecommendationsEnabledSelector);
-  const ff = useFeature("llmNanoSUpsellBanners");
-  const tracking = isOptIn ? "opted_in" : "opted_out";
-  const params = ff?.params?.[tracking];
-
-  const knownDeviceModelIds = useSelector(knownDeviceModelIdsSelector);
-  const hasOnlySeenOneModel = Object.values(knownDeviceModelIds).filter(Boolean).length === 1;
-  const hasOnlySeenLNS = hasOnlySeenOneModel && knownDeviceModelIds.nanoS;
-
-  const accounts = useSelector(accountsSelector);
-  const swapCount = accounts.reduce((count, account) => count + account.swapHistory.length, 0);
-
-  const isEnabled = ff?.enabled && params?.[location];
-  const isShown = Boolean(isEnabled && hasOnlySeenLNS && swapCount < 2);
+  const { isShown, params, tracking } = useLNSUpsellBannerState(location);
 
   const { "%": discount, link: ctaLink } = params ?? {};
   const analitycsPage = AnalyticsPageMap[location];
