+++ apps/ledger-live-mobile/src/newArch/features/LNSUpsell/index.ts
@@ -1,1 +1,2 @@
 export * from "./components/LNSUpsellBanner";
+export * from "./hooks/useLNSUpsellBannerState";
