+++ apps/ledger-live-mobile/src/newArch/features/LNSUpsell/types.ts
@@ -1,0 +1,6 @@
+import type { LlmNanoSUpsellBannersConfig } from "@ledgerhq/types-live/lib/lnsUpsell";
+
+export type LNSBannerLocation = Extract<
+  "manager" | "accounts" | "notification_center" | "wallet",
+  keyof LlmNanoSUpsellBannersConfig
+>;
